requirements.txt
# YOLO & TensorFlow Training Dependencies
ultralytics==8.2.90
torch==2.2.2
torchvision==0.17.2
opencv-python==4.9.0.80
matplotlib==3.8.0
pandas==2.1.0
numpy==1.24.3
tensorboard==2.14.0
Pillow==10.0.0
pyyaml==6.0.1